+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Change log
 - Updated `Krizalys\Onedrive\Proxy\FileProxy`.
 - Updated `Krizalys\Onedrive\Proxy\FolderProxy`.
 - Updated `Krizalys\Onedrive\Proxy\FileSystemInfoProxy`.
+- Updated `Krizalys\Onedrive\Proxy\PackageProxy`.
 
 [2.4.2] - 2019-09-16
 --------------------